requirements.txt
fastapi~= 0.120.1
uvicorn~= 0.38.0
pydantic~= 2.12.3
python-dotenv~= 1.2.1
passlib[bcrypt]~= 4.0.1
pyjwt~=  2.10.1
sqlite3-binary
email-validator~= 2.3.0
secure-smtplib~=  0.1.1

